+++ src/onegov.reservation/onegov/reservation/__init__.py
@@ -1,12 +1,13 @@
-from libres.db.models import Allocation, Reservation
 from onegov.reservation.collection import ResourceCollection
 from onegov.reservation.core import LibresIntegration
+from onegov.reservation.models import PricedAllocation as Allocation
+from onegov.reservation.models import PayableReservation as Reservation
 from onegov.reservation.models import Resource
 
 __all__ = [
     'Allocation',
     'LibresIntegration',
+    'Reservation',
     'Resource',
     'ResourceCollection',
-    'Reservation'
 ]