+++ src/onegov.reservation/onegov/reservation/models/__init__.py
@@ -1,0 +1,5 @@
+from onegov.reservation.models.priced_allocation import PricedAllocation
+from onegov.reservation.models.payable_reservation import PayableReservation
+from onegov.reservation.models.resource import Resource
+
+__all__ = ('PricedAllocation', 'PayableReservation', 'Resource')