+++ src/onegov.reservation/onegov/reservation/models/priced_allocation.py
@@ -1,0 +1,6 @@
+from libres.db.models import Allocation
+from onegov.core.orm import ModelBase
+
+
+class PricedAllocation(Allocation, ModelBase):
+    __mapper_args__ = {'polymorphic_identity': 'priced'}